requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
elasticsearch==8.10.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
spacy==3.7.2
python-multipart==0.0.6
jinja2==3.1.2
google-api-python-client==2.108.0
spacy==3.7.2
numpy==1.26.2
pandas==2.1.3
elasticsearch==8.10.0
together==1.4.0